+++ modules/main/src/main/java/com/nisovin/shopkeepers/compat/NMSManager.java
@@ -34,6 +34,8 @@ public final class NMSManager {
 	// frequently.
 	static {
 		// Registered in the order from latest to oldest.
+		register(new CompatVersion("1_21_R3", "1.21.3", "61a218cda78417b6039da56e08194083"));
+		// Note: 1.21.2 was immediately replaced by 1.21.3 and is not supported.
 		register(new CompatVersion("1_21_R2", "1.21.1", "7092ff1ff9352ad7e2260dc150e6a3ec"));
 		register(new CompatVersion("1_21_R1", "1.21", "229d7afc75b70a6c388337687ac4da1f"));
 		// Note: MC 1.20.6 completely replaced 1.20.5. We only support 1.20.6.