+++ modules/main/src/main/java/com/nisovin/shopkeepers/compat/api/NMSCallProvider.java
@@ -3,6 +3,7 @@ package com.nisovin.shopkeepers.compat.api;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
+import org.bukkit.entity.Salmon;
 import org.bukkit.event.block.BlockExplodeEvent;
 import org.bukkit.event.entity.EntityExplodeEvent;
 import org.bukkit.inventory.ItemStack;
@@ -99,4 +100,10 @@ public interface NMSCallProvider {
 	public boolean isDestroyingBlocks(EntityExplodeEvent event);
 
 	public boolean isDestroyingBlocks(BlockExplodeEvent event);
+
+	// MC 1.21.3+ TODO Can be removed once we only support Bukkit 1.21.3+
+
+	public default void setSalmonVariant(Salmon salmon, String variant) {
+		// Not supported by default.
+	}
 }